refactor: new project structure -> logic and presentation separation
