--- src/main/java/karel/hudera/rnma/characters/IGameCharacter.java
@@ -1,6 +1,0 @@
-package karel.hudera.rnma.characters;
-
-public interface IGameCharacter {
-    String speak();
-    Boolean isAlive();
-}
