+++ src/main/java/karel/hudera/rnma/logic/characters/IGameCharacter.java
@@ -1,0 +1,6 @@
+package karel.hudera.rnma.logic.characters;
+
+public interface IGameCharacter {
+    String speak();
+    Boolean isAlive();
+}
