--- src/main/java/karel/hudera/rnma/logic/commands/ICommand.java
@@ -1,6 +1,0 @@
-package karel.hudera.rnma.logic.commands;
-
-public interface ICommand {
-    String makeCommand(String... param);
-    String getName();
-}
