+++ src/main/java/karel/hudera/rnma/logic/mechanic/commands/ICommand.java
@@ -1,0 +1,6 @@
+package karel.hudera.rnma.logic.mechanic.commands;
+
+public interface ICommand {
+    String makeCommand(String... param);
+    String getName();
+}
