+++ src/main/java/karel/hudera/rnma/logic/mechanic/game/IGame.java
@@ -1,0 +1,10 @@
+package karel.hudera.rnma.logic.mechanic.game;
+
+public interface IGame {
+    String getIntro();
+    String getOutro();
+    String handleInput(String input);
+    boolean gameOver();
+    void setGameOver(boolean gameOver);
+    GamePlan getGamePlan();
+}
