+++ src/test/java/karel/hudera/rnma/logic/commands/CommandTest.java
@@ -1,8 +1,9 @@
 package karel.hudera.rnma.logic.commands;
 
-import karel.hudera.rnma.logic.game.Game;
-import karel.hudera.rnma.logic.game.GamePlan;
-import karel.hudera.rnma.strings.StringResources;
+import karel.hudera.rnma.logic.mechanic.commands.Command;
+import karel.hudera.rnma.logic.mechanic.game.Game;
+import karel.hudera.rnma.logic.mechanic.game.GamePlan;
+import karel.hudera.rnma.logic.strings.StringResources;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
