+++ src/test/java/karel/hudera/rnma/logic/game/GamePlanTest.java
@@ -1,9 +1,11 @@
 package karel.hudera.rnma.logic.game;
 
-import karel.hudera.rnma.strings.StringResources;
-import karel.hudera.rnma.characters.GameCharacter;
-import karel.hudera.rnma.player.Inventory;
-import karel.hudera.rnma.rooms.Room;
+import karel.hudera.rnma.logic.mechanic.game.Game;
+import karel.hudera.rnma.logic.mechanic.game.GamePlan;
+import karel.hudera.rnma.logic.strings.StringResources;
+import karel.hudera.rnma.logic.characters.GameCharacter;
+import karel.hudera.rnma.logic.player.Inventory;
+import karel.hudera.rnma.logic.rooms.Room;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
